New dashboard controller
A new controller have been added to manage the main methods of the menu bar inside the web pages. Also, the 'manage' and 'home_searchs' module have been merged into one called 'database_searchs'.
