+++ app/controllers/user_controller.rb
@@ -1,10 +1,9 @@
 class UserController < ApplicationController
+  layout 'main'
+
   def user_home
     redirect_to graphic_path if user_signed_in?
-    render :layout => true
   end
 
-  def forgot_password
-    render :layout => true
-  end
+  def forgot_password; end
 end
